Move some FRI stuff into the FRI module
I think it would be nice to eventually have separate crates for IOP, FRI, PLONK, etc. This is one step toward that.
