+++ plonky2/src/iop/witness.rs
@@ -5,7 +5,7 @@ use num::{BigUint, FromPrimitive, Zero};
 use plonky2_field::extension_field::{Extendable, FieldExtension};
 use plonky2_field::field_types::Field;
 
-use crate::fri::proof::{FriProof, FriProofTarget};
+use crate::fri::witness_util::set_fri_proof_target;
 use crate::gadgets::arithmetic_u32::U32Target;
 use crate::gadgets::biguint::BigUintTarget;
 use crate::gadgets::nonnative::NonNativeTarget;
@@ -258,69 +258,7 @@ pub trait Witness<F: Field> {
             self.set_extension_target(t, x);
         }
 
-        self.set_fri_proof_target(&proof.opening_proof, &proof_target.opening_proof);
-    }
-
-    /// Set the targets in a `FriProofTarget` to their corresponding values in a `FriProof`.
-    fn set_fri_proof_target<H: AlgebraicHasher<F>, const D: usize>(
-        &mut self,
-        fri_proof: &FriProof<F, H, D>,
-        fri_proof_target: &FriProofTarget<D>,
-    ) where
-        F: RichField + Extendable<D>,
-    {
-        self.set_target(fri_proof_target.pow_witness, fri_proof.pow_witness);
-
-        for (&t, &x) in fri_proof_target
-            .final_poly
-            .0
-            .iter()
-            .zip_eq(&fri_proof.final_poly.coeffs)
-        {
-            self.set_extension_target(t, x);
-        }
-
-        for (t, x) in fri_proof_target
-            .commit_phase_merkle_caps
-            .iter()
-            .zip_eq(&fri_proof.commit_phase_merkle_caps)
-        {
-            self.set_cap_target(t, x);
-        }
-
-        for (qt, q) in fri_proof_target
-            .query_round_proofs
-            .iter()
-            .zip_eq(&fri_proof.query_round_proofs)
-        {
-            for (at, a) in qt
-                .initial_trees_proof
-                .evals_proofs
-                .iter()
-                .zip_eq(&q.initial_trees_proof.evals_proofs)
-            {
-                for (&t, &x) in at.0.iter().zip_eq(&a.0) {
-                    self.set_target(t, x);
-                }
-                for (&t, &x) in at.1.siblings.iter().zip_eq(&a.1.siblings) {
-                    self.set_hash_target(t, x);
-                }
-            }
-
-            for (st, s) in qt.steps.iter().zip_eq(&q.steps) {
-                for (&t, &x) in st.evals.iter().zip_eq(&s.evals) {
-                    self.set_extension_target(t, x);
-                }
-                for (&t, &x) in st
-                    .merkle_proof
-                    .siblings
-                    .iter()
-                    .zip_eq(&s.merkle_proof.siblings)
-                {
-                    self.set_hash_target(t, x);
-                }
-            }
-        }
+        set_fri_proof_target(self, &proof.opening_proof, &proof_target.opening_proof);
     }
 
     fn set_wire(&mut self, wire: Wire, value: F) {
